FIX summary field definitions & translation keys
The summary field is defined back to front - meaning the field title
was "Content.Summary" - while the content of said column in a GridField
contained "raw" HTML (escaped and thus readable to a user).

Along with this fix are a few translation key updates, in order to
be consistent with other Silverstripe CMS 4 modules.